+++ src/components/coctail-details-page/DetailsIngredients.tsx
@@ -1,5 +1,5 @@
 import { Box, Chip, Stack, Typography } from "@mui/material";
-import { CocktailDetails } from "../types/cocktail-details.type";
+import { CocktailDetails } from "../../types/cocktail-details.type";
 
 type Props = {
   details: CocktailDetails | undefined;
@@ -29,16 +29,17 @@ const DetailsIngredients = ({ details }: Props) => {
       <Typography variant="h6" gutterBottom>
         Ingredients:
       </Typography>
-      <Stack direction="row" spacing={1} flexWrap="wrap">
+      <Stack direction="row" spacing={1} useFlexGap flexWrap="wrap">
         {ingredients.map(
           (ingredient, index) =>
             ingredient && (
               <Chip
+                className="ingredient-chip"
                 key={ingredient}
                 label={`${ingredient}${
                   measures[index] ? ` - ${measures[index]}` : ""
                 }`}
-                sx={{ m: '5px', color: "#fff", backgroundColor: "#2f2f2f" }}
+                sx={{ color: "#fff", backgroundColor: "#2f2f2f", mt: "5px" }}
               />
             )
         )}
