+++ src/components/ordinary-page/MediaCard.tsx
@@ -4,9 +4,9 @@ import CardContent from "@mui/material/CardContent";
 import CardMedia from "@mui/material/CardMedia";
 
 import Typography from "@mui/material/Typography";
-import { Cocktail } from "../types/cocktail.type.js";
+import { Cocktail } from "../../types/cocktail.type.js";
 import { CardActionArea } from "@mui/material";
-import NoImage from '../images/NoImage.jpg'
+import NoImage from '../../images/NoImage.jpg'
 
 type Props = {
   cocktail: Cocktail;
