+++ src/components/ordinary-page/MediaCardWrapper.tsx
@@ -1,9 +1,9 @@
 import { useNavigate } from "react-router-dom";
 import MediaCard from "./MediaCard.js";
-import { useAppDispatch } from "../hooks/use-app-dispatch.hook.js";
-import { actions as cocktailDetailsActions } from "../store/cocktail-details/coctail-details.js";
+import { useAppDispatch } from "../../hooks/use-app-dispatch.hook.js";
+import { actions as cocktailDetailsActions } from "../../store/cocktail-details/coctail-details.js";
 
-import { Cocktail } from "../types/cocktail.type.ts";
+import { Cocktail } from "../../types/cocktail.type.ts";
 
 type Props = {
   cocktail: Cocktail;
